One file less
